afgagaga

+++ src/service/implementation/graph_train/maxl_trainer.py
@@ -115,7 +115,7 @@ class MAXLTrainerImpl(Trainer):
             out_nc = self.model.node_classification(z)
             loss_nc = self.criterion( 
                 out_nc.to(self.device)[:, 1],
-                train.y.to(self.device)
+                train.y.to(self.device, dtype=torch.int64)
             )
             
             # Link Prediction
@@ -138,7 +138,7 @@ class MAXLTrainerImpl(Trainer):
             out_lp = self.model.link_prediction(z, edge_label_index).view(-1)
             loss_lp = self.criterion(
                 out_lp.to(self.device), 
-                edge_label.to(self.device)
+                edge_label.to(self.device, dtype=torch.int64)
             )
             
             running_loss_nc = loss_nc.item()
@@ -196,12 +196,12 @@ class MAXLTrainerImpl(Trainer):
             
             val_loss_nc = self.criterion(
                 out_nc[val.val_mask][:, 1].to(self.device),
-                val.y[val.val_mask].to(self.device)
+                val.y[val.val_mask].to(self.device, dtype=torch.int64)
             )
             
             val_loss_lp = self.criterion(
                 out_lp[val.val_mask].to(self.device), 
-                edge_label[val.val_mask].to(self.device)
+                edge_label[val.val_mask].to(self.device, dtype=torch.int64)
             )
             
             
@@ -242,12 +242,12 @@ class MAXLTrainerImpl(Trainer):
             
             test_loss_nc = self.criterion(
                 out_nc[test.test_mask].to(self.device)[:, 1],
-                test.y[test.test_mask].to(self.device)
+                test.y[test.test_mask].to(self.device, dtype=torch.int64)
             )
             
             test_loss_lp = self.criterion(
                 out_lp[test.test_mask].to(self.device), 
-                edge_label[test.test_mask].to(self.device)
+                edge_label[test.test_mask].to(self.device, dtype=torch.int64)
             )
             
             # Compute and update AdaDW loss
